// src/states/model/garantie-reelle.model.ts
export interface GarantieReelModel {
    id?: string;
    terCode?: string;
    creanCode?: string;
    logeCode?: string;
    typgarReelCode?: string;
    gareelLib?: string;
    gareelObjNum?: string;
    gareelModeAcq?: string;
    gareelObjMont?: string;
    gareelDatinscrip?: string;
    gareelDateCtl?: string;
    gareelType?: string;
    gareelNumSerieType?: string;
}